chatgpt code stubs now compile

+++ filesystem.c
@@ -1,6 +1,7 @@
-#include "fat32.h" // Include the corresponding header file
+#include "filesystem.h" // Include the corresponding header file
+#include "common.h" // Include custom data types
 
-#include <stdio.h> // Include necessary standard headers
+// #include <stdio.h> // Include necessary standard headers
 
 int fat32_init(Fat32FileSystem *fs, void *disk) {
     // Implement FAT32 initialization logic here
@@ -14,13 +15,13 @@ int fat32_mount(Fat32FileSystem *fs) {
     return 0; // Return 0 on success
 }
 
-int fat32_read_file(Fat32FileSystem *fs, const char *filename, void *buffer, size_t size, off_t offset) {
+int fat32_read_file(Fat32FileSystem *fs, const char *filename, void *buffer, uint32_t size, int32_t offset) {
     // Implement FAT32 file read logic here
     // Read data from the specified file into the buffer
     return -1; // Return -1 for now as this is a stub
 }
 
-int fat32_write_file(Fat32FileSystem *fs, const char *filename, const void *buffer, size_t size, off_t offset) {
+int fat32_write_file(Fat32FileSystem *fs, const char *filename, const void *buffer, uint32_t size, int32_t offset) {
     // Implement FAT32 file write logic here
     // Write data from the buffer to the specified file
     return -1; // Return -1 for now as this is a stub
